+++ praw/reddit.py
@@ -309,9 +309,11 @@ class Reddit(object):
                     'modaction': models.ModAction,
                     'moderators': models.ModeratorsWidget,
                     'more': models.MoreComments,
+                    'post-flair': models.PostFlairWidget,
                     'stylesheet': models.Stylesheet,
                     'subreddit-rules': models.RulesWidget,
-                    'textarea': models.TextArea}
+                    'textarea': models.TextArea,
+                    'widget': models.Widget}
         for kind, klass in mappings.items():
             self._objector.register(kind, klass)